coffeecounter/requirements.txt
Django==1.7.5
argparse==1.2.1
djangorestframework==3.0.5
wsgiref==0.1.2
